[dynamo][hops] Remove restore_side_effect from the auto flattening codepath
Earlier, restore_side_effects was doing 2 things
1) Restoring side effects
2) If False, it would also expose intermediates to outputs

This was getting confusing with autograd.Function, where for the
backward, we do want to restore side effects but dont want to expose
intermediates as outputs.

So, here we fully get rid of restore_side_effects, and leave it to the
caller of the auto_flattening speculate subgraph to do this manually. We
introduce allow_side_effects flag, which exposes intermediates to
outputs and also sets up side effects to not raise when it sees a
mutation during hop tracing.

ghstack-source-id: 752d680ab3239854f9c109d59a895f0d818e0682
Pull-Request: https://github.com/pytorch/pytorch/pull/169394