[inductor] Fix bug in aten.pow lowering
If a float is the second argument to aten.pow, the original lowering unconditionally attempts to convert it to an int (to check if we can do a clean integer power calculation). Unfortunately, this throws an error when parsing `float("inf")`. Switch to `is_integer()` instead.

Fixes #170937.

ghstack-source-id: ac535f9984ed5fb616d6d651dbd2c6b93b6b561e
Pull-Request: https://github.com/pytorch/pytorch/pull/170960